tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.size.large.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.canvas.host.size.large</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.size.large</h1>
<p class="desc"></p>

<p class="notes">Not sure how reasonable this is, but the spec doesn't say there's an upper limit on the size.
<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var n = 2147483647; // 2^31 - 1, which should be supported by any sensible definition of "long"
  canvas.width = n;
  canvas.height = n;
  _assertSame(canvas.width, n, "canvas.width", "n");
  _assertSame(canvas.height, n, "canvas.height", "n");
  t.done();

});
</script>
